protrainer/requirements.txt
keyboard
pytesseract
Pillow